refactor repository layout
