+++ tests/test_writer.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2021-2024 Northwestern University.
+#
+# galter-subjects-utils is free software; you can redistribute it and/or
+# modify it under the terms of the MIT License; see LICENSE file for more
+# details.
+
+"""Test generic writer functionality."""
+
+from pathlib import Path
+
+from galter_subjects_utils.reader import read_jsonl
+from galter_subjects_utils.writer import write_jsonl
+
+
+def test_write():
+    filepath = Path(__file__).parent / "test_subjects_mesh.jsonl"
+    entries = [
+        {
+            "id": 'D000015',
+            "tags": ["mesh"],
+            "title": {
+                "en": 'Abnormalities, Multiple'
+            }
+        },
+        {
+            "id": 'D000068098',
+            "tags": ["mesh"],
+            "title": {
+                "en": 'Seed Bank'
+            }
+        },
+        {
+            "id": 'D005368',
+            "tags": ["mesh"],
+            "title": {
+                "en": 'Filariasis'
+            }
+        }
+    ]
+
+    write_jsonl(entries, filepath)
+
+    read_entries = list(read_jsonl(filepath))
+    assert entries == read_entries
+
+    filepath.unlink(missing_ok=True)
